+++ src/remotes/zxsimulator/zxsimremote.ts
@@ -58,12 +58,21 @@ export class ZxSimulatorRemote extends DzrpRemote {
 
 	// History info will not occupy a new element but replace the old element
 	// if PC does not change. USed for LDIR, HALT.
-	protected previouslyStoredPCHistory=-1;
+	protected previouslyStoredPCHistory;
 
+	// TBBlue register handling.
+	protected tbblueRegisterSelectValue;
+
+	// Maps function handlers to registers (the key). As key the tbblueRegisterSelectValue is used.
+	protected tbblueRegisterHandler: Map<number, (port: number, value: number) => void>;
 
 	/// Constructor.
 	constructor() {
 		super();
+		// Init
+		this.previouslyStoredPCHistory=-1;
+		this.tbblueRegisterSelectValue=0;
+		this.tbblueRegisterHandler = new Map<number, (port: number, value: number) => void>();
 		// Set decoder
 		Z80Registers.decoder=new Z80RegistersStandardDecoder();
 		this.cpuRunning=false;
@@ -131,6 +140,59 @@ export class ZxSimulatorRemote extends DzrpRemote {
 
 
 	/**
+	 * Selects active port for TBBlue/Next feature configuration.
+	 * See https://wiki.specnext.dev/TBBlue_Register_Select
+	 * The value is just stored, no further action.
+	 * @param port The written port. (0x243B)
+	 * @param value The tbblue register to select.
+	 */
+	protected tbblueRegisterSelect(port: number, value: number) {
+		this.tbblueRegisterSelectValue=value;
+	}
+
+	/**
+	 * Reads and/or writes the selected TBBlue control register.
+	 * See https://wiki.specnext.dev/TBBlue_Register_Access
+	 * Acts according the value and tbblueRegisterSelectValue,
+	 * i.e. calls the mapped fucntion for the selected register.
+	 * At the moment only the memory slot functions are executed.
+	 * @param port The written port. (0x253B)
+	 * @param value The tbblue register to select.
+	 */
+	protected tbblueRegisterAccess(port: number, value: number) {
+		const func=this.tbblueRegisterHandler.get(this.tbblueRegisterSelectValue);
+		if (func)
+			func(port, value);
+	}
+
+
+	/**
+	 * Changes the tbblue slot/bank association for slots 0-7.
+	 * See https://wiki.specnext.dev/Memory_management_slot_0_bank
+	 * tbblueRegisterSelectValue contains the register (0x50-0x57) respectively the
+	 * slot.
+	 * @param port The written port.
+	 * @param value The bank to map.
+	 */
+	protected tbblueMemoryManagementSlots(port: number, value: number) {
+		const slot=this.tbblueRegisterSelectValue&0x07;
+		if (value==0xFF) {
+			// Handle ROM specially
+			if (slot>1)
+				return;	// not allowed
+			// Choose ROM bank according slot
+			if (slot==0)
+				value=0xFE;
+		}
+		else if(value>223)
+			return;	// not existing bank
+
+		// Change the slot/bank
+		this.zxMemory.setSlot(slot, value);
+	}
+
+
+	/**
 	 * Configures the machine.
 	 * Loads the roms and sets up bank switching.
 	 */
@@ -165,16 +227,25 @@ export class ZxSimulatorRemote extends DzrpRemote {
 
 			// "memoryPagingControl"
 			if (Settings.launch.zsim.memoryPagingControl) {
-				// Load the roms
-				try {
-					// Bank switching.
-					this.zxPorts.registerOutPortFunction(0x7FFD, this.zx128BankSwitch.bind(this));
-				}
-				catch (e) {
-					this.emit('warning', e.message);
+				// Bank switching.
+				this.zxPorts.registerOutPortFunction(0x7FFD, this.zx128BankSwitch.bind(this));
+			}
+
+			// TBBlue
+
+			// "tbblueMemoryManagementSlots"
+			if (Settings.launch.zsim.tbblueMemoryManagementSlots) {
+				// Bank switching.
+				for (let tbblueRegister=0x50; tbblueRegister<=0x57; tbblueRegister++) {
+					this.tbblueRegisterHandler.set(tbblueRegister, this.tbblueMemoryManagementSlots.bind(this));
 				}
 			}
 
+			// If any tbblue register is used then enable tbblue ports
+			if (this.tbblueRegisterHandler.size>0) {
+				this.zxPorts.registerOutPortFunction(0x243B, this.tbblueRegisterSelect.bind(this));
+				this.zxPorts.registerOutPortFunction(0x253B, this.tbblueRegisterAccess.bind(this));
+			}
 		}
 		catch (e) {
 			this.emit('warning', e.message);